refactor: setupAccountReducer in a separate file
