+++ packages/frontend/src/redux/createRootReducer.js
@@ -1,30 +1,7 @@
 import { combineReducers } from 'redux';
 
-import { store } from '../';
-import { wallet } from '../utils/wallet';
-import combinedAccountReducers from './combinedAccountReducers';
 import combinedMainReducers from './combinedMainReducers';
-
-const setupAccountReducer = (history) => {
-    const accounts = Object.keys(wallet.accounts);
-    if (!accounts) {
-        return {};
-    }
-
-    return accounts.reduce((x, accountId) => {
-        const reducer = combineReducers(combinedAccountReducers(history));
-        const inicialState = reducer(store?.getState()[accountId], {});
-
-        return ({
-            ...x,
-            [accountId]: (state = inicialState, action) => (
-                (accountId === wallet.accountId)
-                    ? reducer(state, action)
-                    : state
-            )
-        });
-    }, {});
-};
+import setupAccountReducer from './setupAccountReducer';
 
 export default (history) => combineReducers({
     ...combinedMainReducers(history),
